[String] Fix bad "forwarding" reference code in `TStringBuilderAppender<>`.
Old, bad code: Constructor was trying to forward the rvalue reference to the functor despite it not being a forwarding reference because `TFunctor` is a class template parameter.

Brake up the constructor into lvalue and rvalue reference overloads and enforce that the class template parameter `TFunctor` is not a reference.

Update `ConstructStringBuilderAppender<>()` to remove the reference when passing in the functor type template argument.